pyproject.toml
[tool.poetry]
name = "twfy-votes"
version = "0.1.0"
description = "Django app for storing and querying TheyWorkForYou votes"
authors = ["Alex Parsons <alex.parsons@mysociety.org>"]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.2,<5.0"
pydantic = "^2.8.2"
django-compressor = "^4.5.1"
django-libsass = "^0.9"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
duckdb = "^1.0.0"
httpx = "^0.27.2"
ruamel-yaml = "^0.18.6"
rich = ">=12.0.0"
jinja2 = "^3.1.4"
aioduckdb = "^0.1.0"
jinjasql = { git = "https://github.com/yakhu/jinjasql.git", rev = "f8c62d1bea97d0320bb3676f4f83bd2357ccfe55" }
markupsafe = "2.0.1"
tqdm = "^4.66.5"
typer = "^0.12.5"
mysoc-validator = "^0.2.1"
trogon = "^0.5.0"
django-ninja = "^1.3.0"
markdown = "^3.7"
bs4 = "^0.0.2"
django-debug-toolbar = "^4.4.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
pgcli = "^4.1.0"
pandas-stubs = "^2.2.2.240807"
pytest = "^8.3.2"
pytest-django = "^4.9.0"
djlint = "^1.35.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "twfy_votes.settings"
filterwarnings = "ignore::DeprecationWarning"

[tool.djlint]
blank_line_after_tag = "endblock"
blank_line_before_tag = "load,extends,include,block"
format_js = true
format_css = true
custom_blocks = "markdown,switch,case,featureflag,draft"

[tool.ruff]
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",
    # flake8
    "F",
    # isort
    "I",
]
ignore = [
    # line too long, sorted with formatter where it can be
    "E501",
]


[tool.ruff.lint.isort]
known-first-party = ["hub"]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
django = ["django"]
